fix: 修复文字颜色判断逻辑，处理run级别noFill场景
1. 新增run级别样式优先级判断，当检测到a:rPr声明noFill时直接返回透明色
2. 将shapePath.js中的height解析从int改为float，保留小数精度

+++ src/shapePath.js
@@ -6,7 +6,7 @@ import { getTextByPathList } from './utils'
 function shapePie(H, w, adj1, adj2, isClose) {
   const pieVal = parseFloat(adj2)
   const piAngle = parseFloat(adj1)
-  const size = parseInt(H)
+  const size = parseFloat(H)
   const radiusY = size / 2
   const radiusX = w / 2
   const centerX = radiusX
